requirements.txt
opencv-python
imutils
numpy
matplotlib
scikit-learn
pandas
streamlit
altair 
mysql-connector
pyserial